Fix #6: Detect over/under flow

+++ libpostfix/src/main/java/edu/utoledo/nlowe/postfix/BinaryOperator.java
@@ -4,5 +4,5 @@ package edu.utoledo.nlowe.postfix;
  * A functional interface for Binary Operators (those that take two arguments)
  */
 public interface BinaryOperator extends Operator{
-    int evaluate(int a, int b);
+    long evaluate(long a, long b);
 }
