+++ libpostfix/src/main/java/edu/utoledo/nlowe/postfix/UnaryOperator.java
@@ -4,5 +4,5 @@ package edu.utoledo.nlowe.postfix;
  * A functional interface for Binary Operators (those that take one argument)
  */
 public interface UnaryOperator extends Operator {
-    int evaluate(int input);
+    long evaluate(long input);
 }
